@@ -2,10 +2,10 @@
   "name": "serverless-plugin-typescript-example",
   "type": "module",
   "devDependencies": {
-    "@types/aws-lambda": "8.10.108",
-    "serverless": "3.24.1",
-    "serverless-offline": "11.3.0",
-    "serverless-plugin-typescript": "2.1.4",
-    "typescript": "4.8.4"
+    "@types/aws-lambda": "^8.10.121",
+    "serverless": "^3.35.2",
+    "serverless-offline": "^13.0.0",
+    "serverless-plugin-typescript": "^2.1.5",
+    "typescript": "^5.2.2"
   }
 }
