@@ -6,12 +6,12 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@types/aws-lambda": "8.10.108",
-    "serverless": "3.24.1",
-    "serverless-offline": "11.3.0",
-    "serverless-webpack": "5.10.0",
-    "ts-loader": "9.4.1",
-    "typescript": "4.8.4",
-    "webpack": "5.75.0"
+    "@types/aws-lambda": "^8.10.121",
+    "serverless": "^3.35.2",
+    "serverless-offline": "^13.0.0",
+    "serverless-webpack": "^5.13.0",
+    "ts-loader": "^9.4.4",
+    "typescript": "^5.2.2",
+    "webpack": "^5.88.2"
   }
 }
