@@ -6,10 +6,10 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@types/aws-lambda": "8.10.108",
-    "serverless": "3.24.1",
-    "serverless-offline": "11.3.0",
-    "ts-node": "10.9.1",
-    "typescript": "4.8.4"
+    "@types/aws-lambda": "^8.10.121",
+    "serverless": "^3.35.2",
+    "serverless-offline": "^13.0.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
   }
 }
