@@ -1,7 +1,7 @@
 {
   "dedicatedTo": "Blue, a great migrating bird.",
   "name": "serverless-offline",
-  "version": "12.0.4",
+  "version": "13.1.1",
   "description": "Emulate AWS λ and API Gateway locally when developing your Serverless project",
   "license": "MIT",
   "exports": {
@@ -18,6 +18,7 @@
     "prepare-release": "standard-version && prettier --write CHANGELOG.md",
     "prepublishOnly": "npm run lint",
     "prettier": "prettier --check .",
+    "prettier:fix": "prettier --write .",
     "prettier-check": "prettier -c --ignore-path .gitignore \"**/*.{css,html,js,json,md,yaml,yml}\"",
     "prettier-check:updated": "pipe-git-updated --ext=css --ext=html --ext=js --ext=json --ext=md --ext=yaml --ext=yml -- prettier -c",
     "prettify": "prettier --write --ignore-path .gitignore \"**/*.{css,html,js,json,md,yaml,yml}\"",
@@ -51,7 +52,7 @@
   ],
   "author": "David Hérault <dherault@gmail.com> (https://github.com/dherault)",
   "engines": {
-    "node": ">=16.17.0"
+    "node": ">=18.12.0"
   },
   "standard-version": {
     "skip": {
@@ -78,18 +79,18 @@
     ]
   },
   "dependencies": {
-    "@aws-sdk/client-lambda": "^3.414.0",
+    "@aws-sdk/client-lambda": "^3.418.0",
     "@hapi/boom": "^10.0.1",
-    "@hapi/h2o2": "^10.0.3",
+    "@hapi/h2o2": "^10.0.4",
     "@hapi/hapi": "^21.3.2",
-    "@serverless/utils": "^6.8.2",
+    "@serverless/utils": "^6.15.0",
     "array-unflat-js": "^0.1.3",
     "boxen": "^7.1.1",
     "chalk": "^5.3.0",
     "desm": "^1.3.0",
     "execa": "^8.0.1",
     "fs-extra": "^11.1.1",
-    "is-wsl": "^3.0.0",
+    "is-wsl": "^3.1.0",
     "java-invoke-local": "0.0.6",
     "jose": "^4.14.6",
     "js-string-escape": "^1.0.1",
@@ -97,29 +98,28 @@
     "jsonschema": "^1.4.1",
     "jszip": "^3.10.1",
     "luxon": "^3.2.0",
-    "node-fetch": "^3.3.2",
     "node-schedule": "^2.1.1",
     "p-memoize": "^7.1.1",
-    "p-retry": "^6.0.0",
+    "p-retry": "^6.1.0",
     "velocityjs": "^2.0.6",
-    "ws": "^8.14.1"
+    "ws": "^8.14.2"
   },
   "devDependencies": {
     "@istanbuljs/esm-loader-hook": "^0.2.0",
     "archiver": "^6.0.1",
-    "eslint": "^8.49.0",
+    "eslint": "^8.50.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-unicorn": "^46.0.1",
+    "eslint-plugin-unicorn": "^48.0.1",
     "git-list-updated": "^1.2.1",
     "husky": "^8.0.3",
     "lint-staged": "^14.0.1",
     "mocha": "^10.2.0",
     "nyc": "^15.1.0",
     "prettier": "^3.0.3",
-    "serverless": "^3.26.0",
+    "serverless": "^3.35.2",
     "standard-version": "^9.5.0"
   },
   "peerDependencies": {
