@@ -2,8 +2,8 @@
   "name": "serverless-plugin-typescript-test",
   "type": "module",
   "devDependencies": {
-    "@types/aws-lambda": "8.10.109",
-    "serverless-plugin-typescript": "2.1.4",
-    "typescript": "4.9.4"
+    "@types/aws-lambda": "8.10.121",
+    "serverless-plugin-typescript": "2.1.5",
+    "typescript": "5.2.2"
   }
 }
