@@ -1,11 +1,11 @@
 {
   "name": "serverless-webpack-test",
   "devDependencies": {
-    "@types/aws-lambda": "8.10.109",
-    "serverless-webpack": "5.11.0",
-    "ts-loader": "9.4.2",
-    "typescript": "4.9.4",
-    "webpack": "5.75.0"
+    "@types/aws-lambda": "8.10.121",
+    "serverless-webpack": "5.13.0",
+    "ts-loader": "9.4.4",
+    "typescript": "5.2.2",
+    "webpack": "5.88.2"
   },
   "resolutions": {
     "fsevents": "2.1.3"
