@@ -1,6 +1,4 @@
-'use strict'
-
-module.exports = {
+export default {
   rules: {
     'body-leading-blank': [2, 'always'],
     'footer-leading-blank': [2, 'always'],
