@@ -1,8 +1,5 @@
-'use strict'
-
-module.exports = {
+export default {
   endOfLine: 'auto',
   semi: false,
   singleQuote: true,
-  trailingComma: 'all',
 }
