@@ -4,18 +4,16 @@
 // .NET CORE
 // export const supportedDotnetcore = new Set([
 // 'dotnet6',
-// 'dotnetcore3.1',
 // ])
 
 // GO
 export const supportedGo = new Set(['go1.x'])
 
 // JAVA
-export const supportedJava = new Set(['java8', 'java8.al2', 'java11'])
+export const supportedJava = new Set(['java8', 'java8.al2', 'java11', 'java17'])
 
 // NODE.JS
 export const supportedNodejs = new Set([
-  'nodejs12.x',
   'nodejs14.x',
   'nodejs16.x',
   'nodejs18.x',
@@ -26,14 +24,15 @@ export const supportedProvided = new Set(['provided', 'provided.al2'])
 
 // PYTHON
 export const supportedPython = new Set([
-  'python3.6',
   'python3.7',
   'python3.8',
   'python3.9',
+  'python3.10',
+  'python3.11',
 ])
 
 // RUBY
-export const supportedRuby = new Set(['ruby2.7'])
+export const supportedRuby = new Set(['ruby2.7', 'ruby3.2'])
 
 // deprecated runtimes
 // https://docs.aws.amazon.com/lambda/latest/dg/runtime-support-policy.html
