@@ -105,13 +105,9 @@ export default class HttpServer {
           if (request.method === 'options') {
             response.statusCode = 200
 
-            if (request.headers['access-control-expose-headers']) {
-              response.headers['access-control-expose-headers'] =
-                request.headers['access-control-expose-headers']
-            } else {
-              response.headers['access-control-expose-headers'] =
-                'content-type, content-length, etag'
-            }
+            response.headers['access-control-expose-headers'] =
+              request.headers['access-control-expose-headers'] ||
+              'content-type, content-length, etag'
             response.headers['access-control-max-age'] = 60 * 10
 
             if (request.headers['access-control-request-headers']) {
@@ -266,6 +262,7 @@ export default class HttpServer {
 
       log.debug('headers:', headers)
 
+      response.headers = headers
       response.header('Content-Type', 'application/json', {
         duplicate: false,
         override: false,
@@ -296,7 +293,11 @@ export default class HttpServer {
   }
 
   createRoutes(functionKey, albEvent) {
-    const method = albEvent.conditions.method[0].toUpperCase()
+    let method = 'ANY'
+    if ((albEvent.conditions.method || []).length > 0) {
+      method = albEvent.conditions.method[0].toUpperCase()
+    }
+
     const path = albEvent.conditions.path[0]
     const hapiPath = generateAlbHapiPath(path, this.#options, this.#serverless)
 
