@@ -16,7 +16,7 @@ export default class ScheduleEvent {
   source = 'aws.events'
 
   // format of aws displaying the time, e.g.: 2020-02-09T14:13:57Z
-  time = new Date().toISOString().replace(/\.(.*)(?=Z)/g, '')
+  time = new Date().toISOString().replaceAll(/\.(.*)(?=Z)/g, '')
 
   version = '0'
 
