@@ -16,7 +16,7 @@ export default class LambdaFunctionThatReturnsJSONObject {
     },
     service: {
       provider: {
-        runtime: 'nodejs12.x',
+        runtime: 'nodejs18.x',
       },
     },
   }
