@@ -1,37 +1,39 @@
-import assert from 'node:assert'
-import { env } from 'node:process'
-import { join } from 'desm'
-import { setup, teardown } from '../../../_testHelpers/index.js'
-import { BASE_URL } from '../../../config.js'
-import installNpmModules from '../../../installNpmModules.js'
+// import assert from 'node:assert'
+// import { env } from 'node:process'
+// import { join } from 'desm'
+// import { setup, teardown } from '../../../_testHelpers/index.js'
+// import { BASE_URL } from '../../../config.js'
+// import installNpmModules from '../../../installNpmModules.js'
 
 describe('docker and serverless-webpack', function desc() {
-  before(async () => {
-    await installNpmModules(join(import.meta.url, 'app'))
-  })
+  it('placeholder, node.js v12 not supported anymore', () => true)
 
-  beforeEach(async () => {
-    await setup({
-      servicePath: join(import.meta.url, 'app'),
-    })
-  })
+  // before(async () => {
+  //   await installNpmModules(join(import.meta.url, 'app'))
+  // })
 
-  afterEach(() => teardown())
+  // beforeEach(async () => {
+  //   await setup({
+  //     servicePath: join(import.meta.url, 'app'),
+  //   })
+  // })
 
-  it('should work with docker and serverless-webpack', async function it() {
-    // "Could not find 'Docker', skipping tests."
-    if (!env.DOCKER_DETECTED) {
-      this.skip()
-    }
+  // afterEach(() => teardown())
 
-    const url = new URL('/dev/docker-serverless-webpack', BASE_URL)
-    const response = await fetch(url)
-    const json = await response.json()
+  // it('should work with docker and serverless-webpack', async function it() {
+  //   // "Could not find 'Docker', skipping tests."
+  //   if (!env.DOCKER_DETECTED) {
+  //     this.skip()
+  //   }
 
-    const expected = {
-      hello: 'docker and serverless-webpack!',
-    }
+  //   const url = new URL('/dev/docker-serverless-webpack', BASE_URL)
+  //   const response = await fetch(url)
+  //   const json = await response.json()
 
-    assert.deepEqual(json, expected)
-  })
+  //   const expected = {
+  //     hello: 'docker and serverless-webpack!',
+  //   }
+
+  //   assert.deepEqual(json, expected)
+  // })
 })
