@@ -1,5 +1,8 @@
 import { ApolloServer } from '@apollo/server'
-import { startServerAndCreateLambdaHandler } from '@as-integrations/aws-lambda'
+import {
+  handlers,
+  startServerAndCreateLambdaHandler,
+} from '@as-integrations/aws-lambda'
 import gql from 'graphql-tag'
 
 const resolvers = {
@@ -19,4 +22,7 @@ const server = new ApolloServer({
   typeDefs,
 })
 
-export default startServerAndCreateLambdaHandler(server)
+export default startServerAndCreateLambdaHandler(
+  server,
+  handlers.createAPIGatewayProxyEventRequestHandler(),
+)
